==> 04_post-processing-grants.Rmd <==
---
editor_options: 
  chunk_output_type: console
---

# Post-processing of grants 

Here we remove and filter those grants that may have been erroneously classified, following the process of validation and any initial exploratory data analysis
```{r eval=TRUE}
# Load libraries
library(tidyverse)
library(scales)
library(ggsci)
```

```{r, eval=TRUE}
data <- read_csv("data\\classified-grants-for-analysis.csv")
```

Those grants with no funding amount recorded or no keywords from our list contributed nothing to this analysis, and were removed.
```{r}
data <- data %>% 
  filter(Keywords != "['-']") %>%
  drop_na(Grant.Amount) %>%
  dplyr::select(1:42) 
names(data)
```

Here we hard-coded several specific fixes to the classification that repeatedly cropped up during our validation process. All have been checked to have 0 mis-classifications.
```{r}
# Removing grants containing "dinner"
data <- data %>%
  filter(str_detect(Description, "dinner") == FALSE)

# Classifying "dissemination" as act_edu
for(i in 1:nrow(data)) {
  if(str_detect(data$Description[i], "dissemination") == TRUE) {
    data$cat_act_edu[i] = 1
  }
}

# Removing threat_energytransport classification for "palm oil"
for(i in 1:nrow(data)) {
  if(str_detect(data$Description[i], "palm oil|oil-palm|oil palm") == TRUE) {
    data$cat_threat_energytransport[i] = 0
  }
}

# Removing hab_other classification for "aquatic" (will be freshwater/marine)
for(i in 1:nrow(data)) {
  if(str_detect(data$Description[i], "aquatic") == TRUE) {
    data$cat_habitat_other[i] = 0
  }
}
```

Based on advice from the data provider, we only included grants between 2003-2016
```{r}
data <- data %>%
  filter(Year.Authorized >= 2003 & Year.Authorized <= 2016)
```

We ensured that our study focused on in situ conservation grants by eliminating classifications from grants that may be directed toward captive management (aquarium/zoo category).
```{r}
for(i in 1:nrow(data)){
  if(data$cat_aquarium_zoo[i]==1){
    data[i,c(11:17, 19:28, 30:42)]=0
  }
}
```

Here, we built dummy variables that take the sum of our binary classification data across the columns of each broader category.
```{r}
data <- data %>%
  mutate(Num.Class = rowSums(.[,11:length(.)])) %>%
  mutate(Num.Thr = rowSums(.[37:42])) %>%
  mutate(Num.Act = rowSums(.[11:17])) %>%
  mutate(Num.Hab = rowSums(.[c(25,26,28)])) %>%
  mutate(Num.Tax = rowSums(.[30:36])) %>%
  mutate(Num.Bio = rowSums(.[19:20]))
```

We sought to only consider grants that were focused on biodiversity conservation. Upon exploring the data, we found that many grants classified into threat_climate and threat_energytransportation were often not addressing biodiversity. Accordingly, we only included grants classified into these two sub-categories if the descriptions mentioned a biodiversity-related keyword (a habitat, taxa, or general word such as "conservation").

Removing climate/energy grants that do not contain biodiversity conservation keywords:
```{r}
## Category Lists
clim.energy_cat <- names(data)[c(37,38)]                  # climate and energy/transport categories

bio.filter_cat <- names(data)[c(19:28, 30:36)]            # biodiversity, taxa, habitat categories

clim.energy <- data %>%
  filter_at(vars(clim.energy_cat), any_vars(.>0))
clim.energy_bio <- clim.energy %>%
  filter_at(vars(bio.filter_cat), any_vars(.>0))
clim.energy_non.bio <- anti_join(clim.energy, clim.energy_bio)

data_bio <- data %>%
  filter_at(vars(clim.energy_cat), all_vars(.==0)) %>%
  bind_rows(clim.energy_bio)

```

At this point, this is our final dataset of classified grants to analyze. This is our "data_multi" dataset, as grants are classified into all categories/sub-categories that the algorithm classified them into; multiple sub-category classifications are included in this dataframe.
```{r}
data.multi <- data_bio

write_csv(data.multi,"data\\post-processed-grants-for-analysis-multi.csv")
```

In the "data_multi" dataset, grants classified into multiple sub-categories (e.g. threat_climate + threat_habitat) would be over-represented in our analysis of funding. We sought to only count the amount of funding for a grant once within each category. Otherwise,  To accomplish this, we modified the dataset in several ways:  

- Grants classified into two sub-categories were classified into a new column combining the two (e.g. threat_climate_habitat).  
- Grants classified into three or more sub-categories were classified into new "other" column (e.g. threat_other).  
- The habitat category had joint sub-categories in the initial list of categories (e.g. hab_marine_terrestrial). Any grants classified into both a single (e.g. hab_marine) and an overlapping joint sub-category (e.g. hab_marine_terrestrial) were classified as only the joint (e.g. hab_marine_terrestrial).  
  
Full code exists in supplementary code folder. Subsequently, "unique" data file was written and used for further analysis
```{r}
source("code//code_01_unique-grant-classification.R")
write_csv(data.unique, "data\\post-processed-grants-for-analysis-unique.csv")
```
