==> 05-exploratory-data-analysis-funding.Rmd <==
---
editor_options: 
  chunk_output_type: console
---

Here, we will carry out exploratory data analysis of grants to obtain a sense of how money has been allocated across threat and taxonomic groups.

1. Dealing with grants that were classified into multiple categories

The code below governs the way we dealt with multiple, potentially overlapping categories on individual grants. The general format is as follows:  

- If a grant description had two keyword matches with two sub-categories of the same category (e.g. threat_climate & threat_energytransport), it was classified into a joint sub-category (threat_climate_energytransport). 
- Grants that were classified into three or more sub-categories within a category (e.g. threat_climate & threat_energytransport & threat_pollution) were classified into a sub-category labeled as “other” (e.g. threat_other). 


Read in the post-processed grants for analysis
```{r}
dat.1 <- read_csv('data\\post-processed-grants-for-analysis.csv')
```

Here, we built dummy variables that take the sum of our binary classification data across the columns of each broader category.
```{r}
dat.1 <- dat.1 %>%
  mutate(Num.Class = rowSums(.[,11:length(.)])) %>%
  mutate(Num.Thr = rowSums(.[37:42])) %>%
  mutate(Num.Act = rowSums(.[11:17])) %>%
  mutate(Num.Hab = rowSums(.[c(25,26,28)])) %>%
  mutate(Num.Tax = rowSums(.[30:36])) %>%
  mutate(Num.Bio = rowSums(.[19:20]))
```

Multiple threat categories
```{r}
threat <- dat.1[,c(1:10, 37:42)]

threat <- threat %>% 
  mutate(sumAcross = cat_threat_climate + cat_threat_energytransport + 
           cat_threat_habitat + cat_threat_inv_spec + cat_threat_overexp + 
           cat_threat_pollution) %>%
  filter(sumAcross > 0)

# Checking number of grants/money for multiple threat category grants
tot_threat <- sum(threat$Grant.Amount, na.rm = T) 

# Four or more threats retained all grants
threat4 <- threat %>%
  filter(sumAcross <= 4)
nrow(threat4)/nrow(threat)
sum(threat4$Grant.Amount,na.rm = T)/tot_threat


# Three threats retains 99.3% of the money
threat3 <- threat %>%
  filter(sumAcross <= 3)
nrow(threat3)/nrow(threat)
sum(threat3$Grant.Amount,na.rm = T)/tot_threat

# Two threats retained 96% of the money
threat2 <- threat %>%
  filter(sumAcross <= 2)
nrow(threat2)/nrow(threat)
sum(threat2$Grant.Amount, na.rm = T)/tot_threat

# Unique threat categories retained 80.3% of the money
threat1 <- threat %>%
  filter(sumAcross <= 1)
nrow(threat1)/nrow(threat)
sum(threat1$Grant.Amount,na.rm = T)/tot_threat

```

Plot multiple threat categories and grants
```{r plot multi-threats, eval=FALSE}
n_threats <- 1:4
n_threat_grants <- c(nrow(threat1), 
                  nrow(threat2) - nrow(threat1), 
                  nrow(threat3) - nrow(threat2),
                  nrow(threat4) - nrow(threat3))

threats <- data.frame(n_threats, n_threat_grants)

fig_threats_grants <- ggplot(threats, aes(x = n_threats, y = n_threat_grants)) +
  geom_line() + theme_bw() +
  labs(title = "Number of grants across multiple Threat Categories",
       x="Number of threat categories", y="Number of grants")

# save as png
ggsave(fig_threats_grants, filename = "figs/fig_threats_grants.png", 
       height = 4, width = 6, device = png(), dpi = 300); dev.off()

```

```{r export_fig_threats_grants, eval=TRUE, fig.cap="Number of grants that were classified into one or more categories of threat"}
# show exported image
knitr::include_graphics("figs/fig_threats_grants.png")
```

Following the above analysis, we decided to retain grants classified across multiple threat categories for the network analysis. 

Below we visualize overall funding provided by the top 1000 grantmakers (Supplementary Fig S1 - Accumulation of Funding Across Top Grantmakers)
```{r}
# Create donor ranks
donor_Rank <- c(" ", "Top 50", "Top 100", "Top 200", "Top 300", "Top 400", "Top 500", "Top 600", "Top 700", "Top 800", "Top 900", "Top 1000")

# The values below were calculated by summing funding for each of the above donor ranks
sum <- c(0, 14396513202, 16440571398, 18171144637, 18975986509, 19490663536, 19851261916, 20128241008, 20348022885, 20525790876, 20670921174, 20793224389)
sum_mildiv <- sum/1000000
axis_Order <- c(1:12)

sum.data <- data.frame(donor_Rank, sum, sum_mildiv, axis_Order)

fig_topGrantmakers_funding <- ggplot(data = sum.data,
                    mapping = aes(x = axis_Order,
                                  y =  sum_mildiv,
                                  group = 1)) + geom_line() +
  geom_point() +
  geom_area(aes(y=ifelse(sum_mildiv>=0 & sum_mildiv< 19851.26, sum_mildiv, 19851.26)), fill="#696969", alpha = .4) +
  ylim(0, 21500) +
  theme_minimal() +
  xlab("Grantmaker Group") +
  ylab("Sum ($1,000,000USD)") + 
  scale_x_discrete(limits=c(" ", "Top 50", "Top 100", "Top 200", "Top 300", "Top 400", "Top 500", "Top 600", "Top 700", "Top 800", "Top 900", "Top 1000"))

# save as png
ggsave(fig_topGrantmakers_funding, filename = "figs/fig_topGrantmakers_funding.png", 
       height = 4, width = 6, device = png(), dpi = 300); dev.off()

```

```{r export_fig_topGrantmakers_funding, eval=TRUE, fig.cap="Accumulation curve of the funding in our dataset across the top 1000 grantmakers. The shaded region encompasses the amount of funding captured by the top 500 grantmakers, which represents more than 95% of the total amount of funding."}
# show exported image
knitr::include_graphics("figs/fig_topGrantmakers_funding.png")
```
